Fix exceptions from migrations for some setups.
Some commits contain changes to db/schema.rb but don't include the
updated migration-timestamp. This led to the situation in which setups
that started in a certain time-frame want to run some migrations,
which caused exceptions. This patch fixes that situation so the
problematic setups can run all migrations successfully and "catch up".

+++ db/migrate/201508261815_strip_gpg_passphrase.rb
@@ -1,6 +1,8 @@
 class StripGpgPassphrase < ActiveRecord::Migration
   def up
-    remove_column :lists, :gpg_passphrase
+    if column_exists?(:lists, :gpg_passphrase)
+      remove_column :lists, :gpg_passphrase
+    end
   end
 
   def down
